Add Init method to config and sourcers. (#4)
This gets rid of the clunky errorSourcer type that returns errors on all fail-able sourcer methods. There is now an `Init` method that is assumed to be called before any other methods are invoked symmetric to the `Init` method of a nacelle process. This gives us a central place to collect errors that occur on parsing and file operations.