Make dumping source values on error an option, disabled by default. (#19)
Right now, the Config struct will log the source values used to load a
config if it encounters an error while loading. This is great for
troubleshooting but can easily log secrets if using an env sourcer and
running on a platform that sets a number of env secrets by default.

This turns dumping that into an option and keeps it off by default, so
users need to opt-into logging that.